Refactored unit tests by category
